Requested fixes + small fixes
