+++ .gitignore
@@ -2,3 +2,4 @@
 **/dist/       
 **/*.egg-info/
 .idea/
+**/__pycache__/
