+++ Bash_CLI/src/Command/AssignCommand.py
@@ -5,6 +5,6 @@ import src.Controller.Controller as Controller
 
 
 class AssignCommand(Command):
-    def run(self, stdin: str, controller: Controller) -> Tuple[str, str, int]:
+    def run(self, stdin: str, controller: Controller.Controller) -> Tuple[str, str, int]:
         controller.env_vars[self.args[0]] = self.args[1]
         return '', '', 0
