+++ Bash_CLI/src/Command/ExitCommand.py
@@ -5,5 +5,5 @@ import src.Controller.Controller as Controller
 
 
 class ExitCommand(Command):
-    def run(self, stdin: str, controller: Controller) -> Tuple[str, str, int]:
+    def run(self, stdin: str, controller: Controller.Controller) -> Tuple[str, str, int]:
         raise SystemExit
