+++ Bash_CLI/src/Lexer/Lexer.py
@@ -2,7 +2,7 @@ import src.Controller.Controller as Controller
 
 
 class Lexer(object):
-    def __init__(self, controller: Controller):
+    def __init__(self, controller: Controller.Controller):
         self.controller = controller
 
     def parse_to_tokens(self, arg_str):
