+++ Bash_CLI/src/Parser/Parser.py
@@ -2,7 +2,7 @@ import src.Controller.Controller as Controller
 
 
 class Parser(object):
-    def __init__(self, controller: Controller):
+    def __init__(self, controller: Controller.Controller):
         self.controller = controller
 
     def parse_commands(self, list_tokens):
