+++ Bash_CLI/src/Substitutor/Substitutor.py
@@ -2,7 +2,7 @@ import src.Controller.Controller as Controller
 
 
 class Substitutor(object):
-    def __init__(self, controller: Controller):
+    def __init__(self, controller: Controller.Controller):
         self.controller = controller
 
     def resolve_env_var(self, list_tokens):
