+++ Bash_CLI/src/main.py
@@ -1,6 +1,6 @@
-from src.Controller.Controller import Controller
+import src.Controller.Controller as Controller
 
 
 def main():
-    controller = Controller()
+    controller = Controller.Controller()
     controller.run()
